feat: refatorando metodos classes e pacotes, criando classes de servicos de inclusao, alteracoa e delecao do database.
