+++ src/main/java/br/com/itau/cadastrochavepix/domain/deletion/RegistrationDeletion.java
@@ -1,0 +1,8 @@
+package br.com.itau.cadastrochavepix.domain.deletion;
+
+public interface RegistrationDeletion {
+    void deletion();
+
+    void validateEnablePixKey();
+
+}
