+++ src/main/java/br/com/itau/cadastrochavepix/domain/inclusion/RegistrationInclusion.java
@@ -1,0 +1,9 @@
+package br.com.itau.cadastrochavepix.domain.inclusion;
+
+import br.com.itau.cadastrochavepix.model.requests.PixKeyRegisterRequest;
+import br.com.itau.cadastrochavepix.model.responses.PixKeyRegisterResponse;
+
+public interface RegistrationInclusion {
+    public PixKeyRegisterResponse inclusion(PixKeyRegisterRequest request);
+
+}
