+++ src/main/java/br/com/itau/cadastrochavepix/domain/validation/RegistrationValidation.java
@@ -1,0 +1,9 @@
+package br.com.itau.cadastrochavepix.domain.validation;
+
+public interface RegistrationValidation {
+
+    boolean pixKeyValidate(String pixKey);
+
+    boolean accountValidate(String accountType, Integer agencyNumber, Integer accountNumber);
+
+}
