+++ src/main/java/br/com/itau/cadastrochavepix/domain/validation/impl/PhoneValidate.java
@@ -1,0 +1,19 @@
+package br.com.itau.cadastrochavepix.domain.validation.impl;
+
+import br.com.itau.cadastrochavepix.domain.validation.RegistrationValidation;
+
+import java.util.regex.Pattern;
+
+public class PhoneValidate extends AccountValidation implements RegistrationValidation {
+
+    private static Pattern regexPatternPhoneNumber = Pattern.compile("\\+((\\d{11,14}))");
+
+    @Override
+    public boolean pixKeyValidate(String pixKey) {
+        if (pixKey == null) {
+            return false;
+        }
+        return regexPatternPhoneNumber.matcher(pixKey).matches();
+    }
+
+}
