+++ src/main/java/br/com/itau/cadastrochavepix/domain/validation/impl/RandomValidate.java
@@ -1,0 +1,20 @@
+package br.com.itau.cadastrochavepix.domain.validation.impl;
+
+import br.com.itau.cadastrochavepix.domain.validation.RegistrationValidation;
+
+import java.util.regex.Pattern;
+
+public class RandomValidate extends AccountValidation implements RegistrationValidation {
+
+    private static Pattern regexPatternPhoneNumber = Pattern.compile("^[a-zA-Z0-9]{36}$");
+
+    @Override
+    public boolean pixKeyValidate(String pixKey) {
+        if (pixKey == null)
+            return false;
+
+        return regexPatternPhoneNumber.matcher(pixKey).matches();
+    }
+
+
+}
