--- src/main/java/br/com/itau/cadastrochavepix/entity/deletion/PixKeyDeletion.java
@@ -1,8 +1,0 @@
-package br.com.itau.cadastrochavepix.entity.deletion;
-
-public interface PixKeyDeletion {
-    void deletion();
-
-    void validateEnablePixKey();
-
-}
