--- src/main/java/br/com/itau/cadastrochavepix/entity/modification/PixKeyModification.java
@@ -1,29 +1,0 @@
-package br.com.itau.cadastrochavepix.entity.modification;
-
-public class PixKeyModification {
-
-    void modification() {
-
-    }
-
-    void validateAccountType() {
-
-    }
-
-    void validateAgencyNumber() {
-
-    }
-
-    void validateAccountNumber() {
-
-    }
-
-    void validateAccountName() {
-
-    }
-
-    void validateAccountLastName() {
-
-    }
-
-}
