--- src/main/java/br/com/itau/cadastrochavepix/entity/registration/inclusion/PixKeyInclusion.java
@@ -1,9 +1,0 @@
-package br.com.itau.cadastrochavepix.entity.registration.inclusion;
-
-import br.com.itau.cadastrochavepix.model.requests.PixKeyRegisterRequest;
-import br.com.itau.cadastrochavepix.model.responses.PixKeyRegisterResponse;
-
-public interface PixKeyInclusion {
-    public PixKeyRegisterResponse inclusion(PixKeyRegisterRequest request);
-
-}
