--- src/main/java/br/com/itau/cadastrochavepix/entity/registration/validate/RegistrationValidate.java
@@ -1,6 +1,0 @@
-package br.com.itau.cadastrochavepix.entity.registration.validate;
-
-public interface RegistrationValidate {
-    boolean formatValidate(String pixKey);
-
-}
