+++ src/main/java/br/com/itau/cadastrochavepix/model/enums/AccountType.java
@@ -1,0 +1,6 @@
+package br.com.itau.cadastrochavepix.model.enums;
+
+public enum AccountType {
+    CONTA_CORRENTE,
+    CONTA_POUPANCA
+}
