+++ src/main/java/br/com/itau/cadastrochavepix/model/requests/PixKeyDeletionRequest.java
@@ -1,0 +1,7 @@
+package br.com.itau.cadastrochavepix.model.requests;
+
+import lombok.Getter;
+
+public record PixKeyDeletionRequest(
+) {
+}
