+++ src/main/java/br/com/itau/cadastrochavepix/model/requests/PixKeyModificationRequest.java
@@ -1,0 +1,5 @@
+package br.com.itau.cadastrochavepix.model.requests;
+
+public record PixKeyModificationRequest(
+) {
+}
