+++ src/main/java/br/com/itau/cadastrochavepix/model/responses/PixKeyDeletionResponse.java
@@ -1,0 +1,4 @@
+package br.com.itau.cadastrochavepix.model.responses;
+
+public record PixKeyDeletionResponse() {
+}
