+++ src/main/java/br/com/itau/cadastrochavepix/service/database/impl/PixKeyDeletionService.java
@@ -1,0 +1,13 @@
+package br.com.itau.cadastrochavepix.service.database.impl;
+
+import br.com.itau.cadastrochavepix.model.requests.PixKeyDeletionRequest;
+import br.com.itau.cadastrochavepix.model.responses.PixKeyDeletionResponse;
+
+public class PixKeyDeletionService {
+
+    public PixKeyDeletionResponse deletePixKey(PixKeyDeletionRequest pixKeyDeletionRequest) {
+        return null;
+    }
+
+
+}
